refactor: separate styles from app footer component
